Migrate test logic to service
