+++ mediacontroller/src/main/java/com/example/android/mediacontroller/MediaAppTestSuite.kt
@@ -15,40 +15,15 @@
  */
 package com.example.android.mediacontroller
 
-import android.animation.ArgbEvaluator
-import android.app.Dialog
-import android.content.Context
-import android.content.SharedPreferences
-import android.graphics.Color
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
-import android.view.Gravity
-import android.view.View
-import android.view.ViewGroup
-import android.view.LayoutInflater
-import android.view.Window
-import android.widget.LinearLayout
-import android.widget.TextView
-import android.widget.ProgressBar
-import android.widget.Button
-import android.widget.ScrollView
-import androidx.cardview.widget.CardView
-import androidx.core.content.ContextCompat
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.synthetic.main.media_test_option.view.card_header
-import kotlinx.android.synthetic.main.media_test_option.view.card_text
-import kotlinx.android.synthetic.main.media_test_suite_result.view.tests_passing
-import kotlinx.android.synthetic.main.media_test_suite_result.view.tests_passing_header
-import kotlinx.android.synthetic.main.media_test_suite_result.view.total_tests
 
 import java.util.concurrent.Semaphore
 import kotlin.concurrent.thread
 
-
 class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: String, private val testList:
-    Array<TestOptionDetails>, private val testSuiteResultsLayout: RecyclerView, val context: Context) {
+    Array<TestOptionDetails>) {
 
     private val TAG = "MediaAppTestSuite"
 
@@ -56,21 +31,6 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
      * Sleep time after a single test was run. Assures that all steps are flushed out.
      */
     private val SLEEP_TIME = 1000L
-
-    /**
-     * The color that is associated with passing tests.
-     */
-    private val PASSING_COLOR = ContextCompat.getColor(context, R.color.test_result_pass)
-
-    /**
-     * The color that is associated with failing tests.
-     */
-    private val FAILING_COLOR = ContextCompat.getColor(context, R.color.test_result_fail)
-
-    /**
-     * Adapter for displaying individual tests results.
-     */
-    private var resultsAdapter = ResultsAdapter(this.testList)
 
     /**
      * Semaphore to prevent two tests from being run at the same time and interfering with each other.
@@ -98,16 +58,6 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
     private var suiteRunning = false
 
     /**
-     * Object that is used to generate colors for partly passing tests.
-     */
-    private var argbEvaluator = ArgbEvaluator()
-
-    /**
-     * Used to retrieve test configurations.
-     */
-    private val sharedPreferences: SharedPreferences
-
-    /**
      * Thread to handle all of the test suite operations.
      */
     private lateinit var suiteThread: Thread
@@ -118,7 +68,6 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
             positionToIDMap[i] = testList[i].id
             iDToResultsMap[testList[i].id] = TestCaseResults()
         }
-        sharedPreferences = context.getSharedPreferences(MediaAppTestingActivity.SHARED_PREF_KEY_SUITE_CONFIG, Context.MODE_PRIVATE)
     }
 
     /**
@@ -142,40 +91,20 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
     }
 
     /**
-     * Runs a single test suite a specified number of iterations.
-     *
-     * @param numIter - The number of iterations to run the test suite.
+     * Gets all of the tests in the test suite.
      */
-    fun runSuite(numIter: Int) {
-        resetTests()
-        var progressBar :ProgressBar
-        val dialog = Dialog(context)
-        dialog.apply {
-            setCancelable(false)
-            requestWindowFeature(Window.FEATURE_NO_TITLE)
-            setContentView(R.layout.run_suite_iter_dialog)
-            progressBar = findViewById<ProgressBar>(R.id.suite_iter_progress_bar).apply {
-                max = numIter * testList.size
-                progress = -1
-            }
-            findViewById<Button>(R.id.quit_suite_iter_button).setOnClickListener{
-                interrupt()
-                dismiss()
-            }
-            window.setLayout(LinearLayout.LayoutParams.MATCH_PARENT, LinearLayout.LayoutParams.WRAP_CONTENT)
-        }.show()
-        runSuiteImpl(dialog, progressBar, numIter)
+    fun getTestList(): Array<TestOptionDetails> {
+        return testList
     }
 
     /**
      * Runs the full test suite a fixed number of times.
      *
-     * @param dialog - The loading dialog to close when the suite is finished running.
-     * @param progressBar - The progress bar to update with the suites status.
      * @param numIter - The number of times to run the test suite.
      */
-    private fun runSuiteImpl(dialog: Dialog, progressBar: ProgressBar, numIter: Int){
-        resultsAdapter = ResultsAdapter(testList)
+    fun runSuite(numIter: Int, queries: HashMap<String, String>, onStartTest: () -> Unit,
+                 onFinishTestSuite: (idToResultMap: HashMap<Int, TestCaseResults>) -> Unit) {
+        resetTests()
         suiteRunning = true
         suiteThread = thread(start = true) {
             Looper.prepare()
@@ -183,17 +112,16 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
                 for(i in 0 until numIter) {
                     for (test in testList) {
                         resetSingleResults()
-                        progressBar.incrementProgressBy(1)
+                        onStartTest()
 
                         // Flush out any residual media control commands from previous test
                         Thread.sleep(SLEEP_TIME)
 
                         // In the event that a query is not specified, don't run the test.
-                        var query = sharedPreferences.getString(test.name, MediaAppTestingActivity.NO_CONFIG)
+                        var query = MediaAppTestingActivity.NO_CONFIG
+                        query = queries[test.name] ?: MediaAppTestingActivity.NO_CONFIG
                         if (test.queryRequired && query == MediaAppTestingActivity.NO_CONFIG) {
                             test.testResult = TestResult.CONFIG_REQUIRED
-                            val index = positionToIDMap[test.id]
-                            mHandler.post { resultsAdapter.notifyItemChanged(index!!) }
                             continue
                         }
 
@@ -209,10 +137,7 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
             }
             suiteRunning = false
             mHandler.post {
-                Log.i(TAG, "Starting to display results")
-                displayResults()
-                Log.i(TAG, "Finished displaying results")
-                dialog.dismiss()
+                onFinishTestSuite(iDToResultsMap)
             }
         }
     }
@@ -252,7 +177,7 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
     /**
      * Interrupts the test suite and resets its all of its results.
      */
-    private fun interrupt(){
+    fun interrupt(){
         if (!this::suiteThread.isInitialized){
             return
         }
@@ -279,20 +204,8 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
         for (test in testList) {
             iDToResultsMap[test.id] = TestCaseResults()
         }
-        resultsAdapter = ResultsAdapter(arrayOf<TestOptionDetails>())
-        testSuiteResultsLayout.removeAllViews()
-        displayResults()
     }
 
-
-    /**
-     * Displays the test suite results using a RecyclerView.
-     */
-    private fun displayResults() {
-        testSuiteResultsLayout.layoutManager = LinearLayoutManager(context)
-        testSuiteResultsLayout.setHasFixedSize(true)
-        testSuiteResultsLayout.adapter = resultsAdapter
-    }
 
     /**
      * Class to store all of a tests resulting info.
@@ -302,110 +215,5 @@ class MediaAppTestSuite(val testSuiteName: String, val testSuiteDescription: Str
         var numPassing = 0
         var passingLogs = arrayListOf<ArrayList<String>>()
         var failingLogs = arrayListOf<ArrayList<String>>()
-    }
-
-    /**
-     *  Adapter to display each test and the number of associated successes.
-     */
-    inner class ResultsAdapter(
-            private val tests: Array<TestOptionDetails>
-    ) : RecyclerView.Adapter<ResultsAdapter.ViewHolder>() {
-        inner class ViewHolder(val cardView: CardView) : RecyclerView.ViewHolder(cardView)
-
-        override fun onCreateViewHolder(
-                parent: ViewGroup,
-                viewType: Int
-        ): ResultsAdapter.ViewHolder {
-            val cardView = LayoutInflater.from(parent.context)
-                    .inflate(R.layout.media_test_suite_result, parent, false) as CardView
-            return ViewHolder(cardView)
-        }
-
-        override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            val testID = positionToIDMap[position]!!
-            val testCaseResults = iDToResultsMap[testID]!!
-            holder.cardView.card_header.text = tests[position].name
-            holder.cardView.card_text.text = tests[position].desc
-            holder.cardView.total_tests.text = testCaseResults.totalRuns.toString()
-            holder.cardView.tests_passing.text = testCaseResults.numPassing.toString()
-            if(testCaseResults.totalRuns == 0){
-                holder.cardView.setOnClickListener{}
-                holder.cardView.tests_passing_header.text = context.getString(R.string.test_suite_config_needed_header)
-                holder.cardView.setCardBackgroundColor(Color.GRAY)
-                return
-            }
-            val passPercentage = testCaseResults.numPassing.toFloat() / testCaseResults.totalRuns.toFloat()
-            holder.cardView.setCardBackgroundColor((argbEvaluator.evaluate(passPercentage, FAILING_COLOR, PASSING_COLOR )) as Int)
-            val onResultsClickedListener = OnResultsClickedListener(testID, tests[position].name, tests[position].desc, this@MediaAppTestSuite.context)
-            holder.cardView.setOnClickListener(onResultsClickedListener)
-        }
-
-        override fun getItemCount() = tests.size
-    }
-
-    /**
-     *  Listener for when a specific test result is clicked from the results adapter. Shows logs
-     * for runs associated with the specific tests.
-     */
-    inner class OnResultsClickedListener(private val testId: Int, private val name: String, private val description: String, val context: Context) : View.OnClickListener {
-
-        override fun onClick(p0: View?) {
-            var dialog = Dialog(context).apply {
-                requestWindowFeature(Window.FEATURE_NO_TITLE)
-                setContentView(R.layout.test_suite_results_dialog)
-                findViewById<TextView>(R.id.results_title).text = name
-                findViewById<TextView>(R.id.results_subtitle).text = description
-
-                // Add the passing text log section
-                if(iDToResultsMap[testId]!!.passingLogs.isNotEmpty()) {
-                    val passing_results_log = findViewById<LinearLayout>(R.id.passing_results_log)
-                    passing_results_log.removeAllViews()
-                    for (logsList in iDToResultsMap[testId]!!.passingLogs) {
-                        passing_results_log.addView(TextView(context).apply {
-                            text = resources.getString(R.string.test_iter_divider)
-                            setTextAppearance(context, R.style.SubHeader)
-                            gravity = Gravity.CENTER
-                            setTextColor(PASSING_COLOR)
-                        })
-                        for (line in logsList) {
-                            var logLine = TextView(context).apply {
-                                text = line
-
-                                setTextAppearance(context, R.style.SubText)
-                            }
-                            passing_results_log.addView(logLine)
-                        }
-                    }
-                } else {
-                    findViewById<TextView>(R.id.passing_logs_header).visibility = View.GONE
-                }
-
-                // Add the passing text log section
-                if(iDToResultsMap[testId]!!.failingLogs.isNotEmpty()) {
-                    val failing_results_log = findViewById<LinearLayout>(R.id.failing_results_log)
-                    failing_results_log.removeAllViews()
-                    for (logsList in iDToResultsMap[testId]!!.failingLogs) {
-                        failing_results_log.addView(TextView(context).apply {
-                            text = resources.getString(R.string.test_iter_divider)
-                            setTextAppearance(context, R.style.SubHeader)
-                            gravity = Gravity.CENTER
-                            setTextColor(FAILING_COLOR)
-                        })
-                        for (line in logsList) {
-                            var logLine = TextView(context).apply {
-                                text = line
-                                setTextAppearance(context, R.style.SubText)
-                            }
-                            failing_results_log.addView(logLine)
-                        }
-                    }
-                }
-                else {
-                    findViewById<TextView>(R.id.failing_logs_header).visibility = View.GONE
-                }
-                findViewById<ScrollView>(R.id.results_scroll_view).layoutParams.height = (MediaAppTestingActivity.getScreenHeightPx(context) / 2).toInt()
-                findViewById<Button>(R.id.close_results_button).setOnClickListener(View.OnClickListener { dismiss() })
-            }.show()
-        }
     }
 }
